requirements.txt
aiohttp==3.7.3
Flask==1.1.2
Flask-Dropzone==1.5.4
Flask_RESTful==0.3.8
Flask-Session==0.3.2
gunicorn==20.0.4
numpy==1.17.4
pandas==1.0.5
psycopg2_binary==2.8.6
python-dotenv==0.15.0
yoyo_migrations==7.2.0
testing.postgresql==1.3.0
tqdm==4.46.0
WTForms==2.3.3
https://download.pytorch.org/whl/cpu/torch-1.7.1%2Bcpu-cp37-cp37m-linux_x86_64.whl
